created Admin
Admin's model, repository, dto and service

+++ backend/src/main/java/com/authenticket/authenticket/dto/event/EventDto.java
@@ -3,7 +3,7 @@ package com.authenticket.authenticket.dto.event;
 import java.time.LocalDateTime;
 
 //currently only hides the deleted at field
-public record EventDto(Long eventId,
+public record EventDto(Integer eventId,
                        String eventName,
                        String eventDescription,
                        LocalDateTime eventDate,
